fix: update date, delete hackathon

+++ backend/src/hackathon/infrastructure/database/repositories/mongoose-hackathon.repository.ts
@@ -2,7 +2,7 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Hackathon } from 'src/hackathon/domain/entities/hackathon.entity';
 import { HackathonRepository } from 'src/hackathon/domain/repositories/hackathon.repository';
 import { HackathonDocument } from '../schemas';
-import { Model } from 'mongoose';
+import { Model, Types } from 'mongoose';
 import { NotFoundException } from '@nestjs/common';
 import { UserDocument } from 'src/user/infrastructure/database/schemas';
 import { ProjectDocument } from 'src/project/infrastructure/database/schemas';
@@ -118,14 +118,25 @@ export class MongooseHackathonRepository implements HackathonRepository {
     return updatedHackathon;
   }
 
-  async delete(id: string): Promise<string> {
+  async delete(userId: string, id: string): Promise<string> {
     const existingHackathon = await this.hackathonModel.findById(id);
 
     if (!existingHackathon) {
       throw new NotFoundException(`Hackathon with ID ${id} not found.`);
     }
 
+    const existingUser = await this.userModel.findById(userId);
+
+    if (!existingUser) {
+      throw new NotFoundException(`User with ID ${userId} not found.`);
+    }
+
     await this.hackathonModel.findByIdAndDelete(id);
+    await this.userModel.findByIdAndUpdate(
+      userId,
+      { $pull: { hackathons: new Types.ObjectId(id) } },
+      { new: true },
+    );
 
     return 'Delete successfully';
   }
